+++ packages/pn-commons/src/index.tsx
@@ -12,7 +12,10 @@ import StatusTooltip from './components/Notifications/StatusTooltip';
 import NotificationsCard from './components/Notifications/NotificationsCard';
 import OutlinedButton from './components/OutlinedButton';
 import TitleAndDescription from './components/TitleAndDescription';
-import CustomMobileDialog from './components/CustomMobileDialog';
+import CustomMobileDialog from './components/CustomMobileDialog/CustomMobileDialog';
+import CustomMobileDialogToggle from './components/CustomMobileDialog/CustomMobileDialogToggle';
+import CustomMobileDialogContent from './components/CustomMobileDialog/CustomMobileDialogContent';
+import CustomMobileDialogAction from './components/CustomMobileDialog/CustomMobileDialogAction';
 
 export { LoadingOverlay };
 export { Header };
@@ -28,6 +31,9 @@ export { NotificationsCard };
 export { OutlinedButton };
 export { TitleAndDescription };
 export { CustomMobileDialog };
+export { CustomMobileDialogToggle };
+export { CustomMobileDialogContent };
+export { CustomMobileDialogAction };
 
 // pages
 import NotFound from './navigation/NotFound';
