--- packages/pn-commons/src/types/CustomMobileDialog.ts
@@ -1,7 +1,0 @@
-import { ReactNode } from 'react';
-
-export type CustomDialogAction = {
-  key: string,
-  component: ReactNode,
-  closeOnClick?: boolean
-}
